feat(react): add format option for buildable libs
- Clean up @nrwl/web:package executor
- Combine react package e2e tests to speed up test time
