+++ docs/react/api-web/executors/package.md
@@ -77,6 +77,14 @@ Type: `boolean`
 
 CSS files will be extracted to the output folder.
 
+### format
+
+Alias(es): f
+
+Type: `array`
+
+Only build the specified comma-separated formats (esm,umd,cjs)
+
 ### globals
 
 Type: `object[]`
