+++ package.json
@@ -77,11 +77,11 @@
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
     "@popperjs/core": "^2.9.2",
     "@reduxjs/toolkit": "1.5.0",
-    "@rollup/plugin-babel": "5.0.2",
-    "@rollup/plugin-commonjs": "11.0.2",
-    "@rollup/plugin-image": "2.0.4",
+    "@rollup/plugin-commonjs": "^20.0.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-image": "^2.1.0",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "7.1.1",
+    "@rollup/plugin-node-resolve": "^13.0.4",
     "@schematics/angular": "~12.1.0",
     "@storybook/addon-essentials": "~6.3.0",
     "@storybook/addon-knobs": "~6.3.0",
@@ -114,7 +114,7 @@
     "angular": "1.8.0",
     "autoprefixer": "^10.2.5",
     "babel-jest": "27.0.6",
-    "caniuse-lite": "^1.0.30001030",
+    "caniuse-lite": "^1.0.30001251",
     "chalk": "4.1.0",
     "chokidar": "^3.5.1",
     "circular-dependency-plugin": "5.2.0",
@@ -198,12 +198,11 @@
     "react-router-dom": "5.1.2",
     "regenerator-runtime": "0.13.7",
     "release-it": "^14.11.3",
-    "rollup": "1.31.1",
-    "rollup-plugin-copy": "3.3.0",
-    "rollup-plugin-filesize": "^9.0.0",
-    "rollup-plugin-local-resolve": "1.0.7",
-    "rollup-plugin-peer-deps-external": "2.2.2",
-    "rollup-plugin-postcss": "^4.0.0",
+    "rollup": "^2.56.2",
+    "rollup-plugin-copy": "^3.4.0",
+    "rollup-plugin-filesize": "^9.1.1",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-postcss": "^4.0.1",
     "rollup-plugin-typescript2": "^0.30.0",
     "rxjs": "6.6.7",
     "rxjs-for-await": "0.0.2",
