+++ packages/web/package.json
@@ -41,11 +41,11 @@
     "@babel/plugin-transform-runtime": "^7.15.0",
     "@babel/runtime": "^7.14.8",
     "@babel/preset-typescript": "^7.15.0",
-    "@rollup/plugin-commonjs": "11.0.2",
-    "@rollup/plugin-babel": "5.0.2",
-    "@rollup/plugin-image": "2.0.4",
+    "@rollup/plugin-commonjs": "^20.0.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-image": "^2.1.0",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "7.1.1",
+    "@rollup/plugin-node-resolve": "^13.0.4",
     "autoprefixer": "^10.2.5",
     "babel-loader": "8.1.0",
     "babel-plugin-const-enum": "^1.0.1",
@@ -53,7 +53,7 @@
     "babel-plugin-transform-async-to-promises": "^0.8.15",
     "babel-plugin-transform-typescript-metadata": "^0.3.1",
     "browserslist": "^4.16.6",
-    "caniuse-lite": "^1.0.30001030",
+    "caniuse-lite": "^1.0.30001251",
     "chalk": "4.1.0",
     "chokidar": "^3.5.1",
     "circular-dependency-plugin": "5.2.0",
@@ -80,12 +80,11 @@
     "rxjs": "^6.5.4",
     "rxjs-for-await": "0.0.2",
     "rimraf": "^3.0.2",
-    "rollup": "1.31.1",
-    "rollup-plugin-copy": "^3.3.0",
-    "rollup-plugin-filesize": "^9.0.0",
-    "rollup-plugin-local-resolve": "^1.0.7",
-    "rollup-plugin-peer-deps-external": "^2.2.2",
-    "rollup-plugin-postcss": "^4.0.0",
+    "rollup": "^2.56.2",
+    "rollup-plugin-copy": "^3.4.0",
+    "rollup-plugin-filesize": "^9.1.1",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-postcss": "^4.0.1",
     "rollup-plugin-typescript2": "^0.30.0",
     "sass": "^1.26.3",
     "sass-loader": "8.0.2",
