+++ packages/web/src/executors/package/schema.d.ts
@@ -1,0 +1,23 @@
+export interface Globals {
+  moduleId: string;
+  global: string;
+}
+
+export interface WebPackageOptions {
+  outputPath: string;
+  tsConfig: string;
+  project: string;
+  entryFile: string;
+  extractCss?: boolean;
+  globals?: Globals[];
+  external?: string[];
+  rollupConfig?: string | string[];
+  babelConfig?: string;
+  watch?: boolean;
+  assets?: any[];
+  updateBuildableProjectDepsInPackageJson?: boolean;
+  buildableProjectDepsInPackageJsonType?: 'dependencies' | 'peerDependencies';
+  umdName?: string;
+  deleteOutputPath?: boolean;
+  format: string[];
+}
