+++ packages/web/src/utils/types.ts
@@ -39,29 +39,6 @@ export interface BuildBuilderOptions {
   sourceRoot?: string;
 }
 
-export interface Globals {
-  moduleId: string;
-  global: string;
-}
-
-export interface PackageBuilderOptions {
-  outputPath: string;
-  tsConfig: string;
-  project: string;
-  entryFile: string;
-  extractCss?: boolean;
-  globals?: Globals[];
-  external?: string[];
-  rollupConfig?: string | string[];
-  babelConfig?: string;
-  watch?: boolean;
-  assets?: any[];
-  updateBuildableProjectDepsInPackageJson?: boolean;
-  buildableProjectDepsInPackageJsonType?: 'dependencies' | 'peerDependencies';
-  umdName?: string;
-  deleteOutputPath?: boolean;
-}
-
 export interface AssetGlobPattern {
   glob: string;
   input: string;
